Fixed bugs that happen in the rare cases when the entire dataset is filtered out

+++ clust/scripts/io.py
@@ -321,5 +321,6 @@ def getparallelprogress():
 
 
 def deletetmpfile():
-    os.remove(glob.tmpfile)
+    if os._exists(glob.tmpfile):
+        os.remove(glob.tmpfile)
